+++ Siv3D/Source/Siv3D/Sprite/SivSprite.cpp
@@ -17,7 +17,7 @@ namespace s3d
 {
 	void Sprite::draw() const
 	{
-		Siv3DEngine::GetRenderer2D()->addSprite(none, *this, 0, indices.size());
+		Siv3DEngine::GetRenderer2D()->addSprite(none, *this, 0, static_cast<uint32>(indices.size()));
 	}
 
 	void Sprite::draw(const Texture& texture) const
